+++ statham/dsl/elements/__init__.py
@@ -8,7 +8,6 @@ from statham.dsl.elements.composition import (
     Not,
     OneOf,
 )
-from statham.dsl.elements.meta import ObjectOptions
 from statham.dsl.elements.null import Null
 from statham.dsl.elements.numeric import Integer, Number
 from statham.dsl.elements.object import Object