+++ statham/dsl/elements/meta.py
@@ -1,8 +1,7 @@
 import keyword
-from typing import Any, Dict, List, Optional, Tuple, Type, Union
+from typing import Any, Dict, List, Tuple, Type, Union
 
 from statham.dsl.elements.base import Element
-from statham.dsl.helpers import custom_repr
 from statham.dsl.property import _Property
 from statham.dsl.constants import NotPassed
 from statham.dsl.exceptions import SchemaDefinitionError
@@ -14,28 +13,8 @@ from statham.dsl.validation import (
 )
 
 
-class ObjectOptions:
-
-    additionalProperties: Optional[Union[Element, bool]]
-
-    def __init__(self, *, additionalProperties: Union[Element, bool] = True):
-        # Name used to match JSONSchema.
-        # pylint: disable=invalid-name
-        self.additionalProperties = additionalProperties
-
-    def __eq__(self, other) -> bool:
-        if not isinstance(other, ObjectOptions):
-            return False
-        return self.additionalProperties == other.additionalProperties
-
-    def __repr__(self):
-        return custom_repr(self)
-
-
 RESERVED_PROPERTIES = (
-    dir(object)
-    + list(keyword.kwlist)
-    + ["default", "options", "properties", "additional_properties"]
+    dir(object) + list(keyword.kwlist) + ["default", "properties", "_dict"]
 )
 
 
@@ -51,16 +30,13 @@ class ObjectClassDict(dict):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.properties = {}
-        self.default = self.get("default", NotPassed())
-        self.options = self.get("options", ObjectOptions())
+        self.default = self.pop("default", NotPassed())
 
     def __setitem__(self, key, value):
         if key in RESERVED_PROPERTIES and isinstance(value, _Property):
             raise SchemaDefinitionError.reserved_attribute(key)
         if key == "default":
             self.default = value
-        if key == "options":
-            self.options = value
         if isinstance(value, _Property):
             value.bind_name(key)
             return self.properties.__setitem__(key, value)
@@ -75,7 +51,7 @@ class ObjectMeta(type, Element):
     """
 
     properties: Dict[str, _Property]
-    options: ObjectOptions
+    additionalProperties: Union[Element, bool]
 
     @staticmethod
     def __subclasses__():
@@ -86,16 +62,18 @@ class ObjectMeta(type, Element):
         return []
 
     @classmethod
-    def __prepare__(mcs, _name, _bases):
+    def __prepare__(mcs, _name, _bases, **_kwargs):
         return ObjectClassDict()
 
-    def __new__(mcs, name: str, bases: Tuple[Type], classdict: ObjectClassDict):
+    def __new__(
+        mcs, name: str, bases: Tuple[Type], classdict: ObjectClassDict, **kwargs
+    ):
         cls: Type[ObjectMeta] = type.__new__(mcs, name, bases, dict(classdict))
         cls.properties = classdict.properties
         for property_ in cls.properties.values():
             property_.bind_class(cls)
         cls.default = classdict.default
-        cls.options = classdict.options
+        cls.additionalProperties = kwargs.get("additionalProperties", True)
         return cls
 
     def __hash__(cls):
@@ -125,29 +103,18 @@ class ObjectMeta(type, Element):
             ),
             AdditionalProperties(
                 {prop.source for prop in cls.properties.values()},
-                cls.options.additionalProperties,
+                cls.additionalProperties,
             ),
         ]
 
-    def construct_additional(cls, name: str, value: Any) -> Any:
-        """Construct properties not specified in `properties`."""
-        element: Element = Element()
-        if isinstance(cls.options.additionalProperties, Element):
-            element = cls.options.additionalProperties
-        additional_property = _Property(element)
-        additional_property.bind_class(cls)
-        additional_property.bind_name(name)
-        return additional_property(value)
-
     def python(cls) -> str:
         super_cls = next(iter(cls.mro()[1:]))
-        class_def = f"""class {repr(cls)}({super_cls.__name__}):
+        cls_args = [super_cls.__name__]
+        if cls.additionalProperties not in (True, Element()):
+            cls_args.append(f"additionalProperties={cls.additionalProperties}")
+        class_def = f"""class {repr(cls)}({', '.join(cls_args)}):
 """
-        if (
-            not cls.properties
-            and isinstance(cls.default, NotPassed)
-            and cls.options == ObjectOptions()
-        ):
+        if not cls.properties and isinstance(cls.default, NotPassed):
             class_def = (
                 class_def
                 + """
@@ -159,12 +126,6 @@ class ObjectMeta(type, Element):
                 class_def
                 + f"""
     default = {repr(cls.default)}
-"""
-            )
-        if cls.options != ObjectOptions():
-            class_def = class_def + (
-                f"""
-    options = {repr(cls.options)}
 """
             )
         for property_ in cls.properties.values():