+++ statham/serializer.py
@@ -15,7 +15,6 @@ from statham.dsl.elements import (
     Number,
     OneOf,
     Object,
-    ObjectOptions,
     String,
 )
 from statham.dsl.property import Property