+++ tests/models/autoname.py
@@ -10,36 +10,27 @@ from statham.dsl.elements import (
     Number,
     OneOf,
     Object,
-    ObjectOptions,
     String,
 )
 from statham.dsl.property import Property
 
 
-class ListOfStringsItem(Object):
-
-    options = ObjectOptions(additionalProperties=False)
+class ListOfStringsItem(Object, additionalProperties=False):
 
     string_property: Maybe[str] = Property(String())
 
 
-class ListOfIntegersItem(Object):
-
-    options = ObjectOptions(additionalProperties=False)
+class ListOfIntegersItem(Object, additionalProperties=False):
 
     integer_property: Maybe[int] = Property(Integer())
 
 
-class ListAnyOfItem0(Object):
-
-    options = ObjectOptions(additionalProperties=False)
+class ListAnyOfItem0(Object, additionalProperties=False):
 
     string_prop: Maybe[str] = Property(String())
 
 
-class ListAnyOfItem1(Object):
-
-    options = ObjectOptions(additionalProperties=False)
+class ListAnyOfItem1(Object, additionalProperties=False):
 
     integer_prop: Maybe[int] = Property(Integer())